+++ CHANGES.rst
@@ -28,6 +28,11 @@ vizier
 Infrastructure, Utility and Other Changes and Additions
 -------------------------------------------------------
 
+- Versions of astropy <5.0 and numpy <1.20 are no longer supported. [#2966]
+
+- Versions of Python <3.9 are no longer supported. [#2966]
+
+
 utils.tap
 ^^^^^^^^^